pyproject.toml
[project]
name = "harina-v4"
version = "0.2.0"
description = "HARINA V4 CLI and Discord receipt bot powered by Gemini, Google Drive, and Google Sheets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "discord.py>=2.6.3",
  "google-api-python-client>=2.175.0",
  "google-auth>=2.41.1",
  "google-genai>=1.30.0",
  "pydantic>=2.11.3",
  "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
]

[project.scripts]
harina = "app.cli:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
